+++ CHANGELOG.md
@@ -32,6 +32,8 @@
 * Non-integer nodeids no longer break sorting and construction
 * Fix a bug in EDS getting CVARSORT when properties are printed
 * Use target relations for writing empty files in `mkprof` command (#125)
+* Properly split role and post in Dmrs.from_triples()
+* Use top-specifying triples (e.g. (0, 'top', 10000)) in Dmrs.from_triples()
 
 ### Removed
 ### Deprecated